requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
groq==0.4.2
langchain==0.0.350
langchain-groq==0.0.1
sentence-transformers==2.2.2
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
langdetect==1.0.9
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1 